Add test cases for `crate::crypto::Address`
* Create a separate directory for the module `crate::crypto`.
* Move `crate::crypto::Address` implementation to a separate file.
* Add test cases for `crate::crypto::Address`.
