+++ src/crypto/mod.rs
@@ -8,43 +8,8 @@ type ChecksumAlg = sha2::Sha512Trunc256;
 const CHECKSUM_LEN: usize = 4;
 const HASH_LEN: usize = 32;
 
-/// Public key address
-#[derive(Debug, Copy, Clone, Eq, PartialEq)]
-pub struct Address(pub [u8; HASH_LEN]);
-
-impl Address {
-    pub fn new(bytes: [u8; HASH_LEN]) -> Address {
-        Address(bytes)
-    }
-
-    /// Decode address from base64 string with checksum
-    pub fn from_string(string: &str) -> Result<Address, String> {
-        let checksum_address = match BASE32_NOPAD.decode(string.as_bytes()) {
-            Ok(decoded) => decoded,
-            Err(err) => return Err(format!("Error decoding base32: {:?}", err)),
-        };
-        if checksum_address.len() != (HASH_LEN + CHECKSUM_LEN) {
-            return Err("Input string is an invalid address. Wrong length".to_string());
-        }
-        let (address, checksum) = checksum_address.split_at(HASH_LEN);
-        let hashed = ChecksumAlg::digest(&address);
-        if &hashed[(HASH_LEN - CHECKSUM_LEN)..] == checksum {
-            let mut bytes = [0; HASH_LEN];
-            bytes.copy_from_slice(address);
-            Ok(Address::new(bytes))
-        } else {
-            Err("Input checksum did not validate".to_string())
-        }
-    }
-
-    /// Encode address to base64 string with checksum
-    pub fn encode_string(&self) -> String {
-        let hashed = ChecksumAlg::digest(&self.0);
-        let checksum = &hashed[(HASH_LEN - CHECKSUM_LEN)..];
-        let checksum_address = [&self.0, checksum].concat();
-        BASE32_NOPAD.encode(&checksum_address)
-    }
-}
+mod address;
+pub use address::Address;
 
 /// Convenience struct for handling multisig public identities
 #[derive(Debug, Clone)]
